refactoring for incoming features
